requirements.txt
discord.py==1.7.3
youtube-dl==2021.6.6
PyNaCl==1.4.0
python-dotenv==0.19.0